+++ include/Histogram.h
@@ -9,19 +9,19 @@
 #include "stdlib.h"
 
 struct Pair {
-    int first;
-    int second;
+    u_int64_t first;
+    u_int64_t second;
 };
 
 typedef struct Histogram {
     struct Pair *hist;
-    unsigned int size;
+    size_t size;
 } Histogram;
 
-    Histogram *Histogram_new(unsigned int size);
+    Histogram *Histogram_new(size_t size);
     void Histogram_free(Histogram *histogram);
-    void Histogram_insert(Histogram *histogram, int i);
-    int Histogram_getElemAtIndex(Histogram *histogram, int i);
+    void Histogram_insert(Histogram *histogram, u_int64_t i);
+    u_int64_t Histogram_getElemAtIndex(Histogram *histogram, u_int64_t i);
     void Histogram_outAsJSON(Histogram *histogram, FILE *out);
 
 #endif //SPARSEMATRIX_HISTOGRAM_H
